requirements.txt
flask==3.0.3
flask-cors==4.0.1
PyMuPDF==1.24.5
pdfplumber==0.11.1
pytesseract==0.3.10
Pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5
rapidfuzz==3.9.7
python-dateutil==2.9.0.post0
